+++ types/Queue.d.ts
@@ -36,6 +36,25 @@ export interface QueueState {
   messages?: MessageMessage[];
 }
 
+export const enum QueueEventNames {
+  /** Consumer was cancelled */
+  QueueConsumerCancel = 'consumer.cancel',
+  /** Consumer was added */
+  QueueConsume = 'consume',
+  /** Message was dead-lettered, sends deadLetterExchange name and message */
+  QueueDeadLetter = 'dead-letter',
+  /** Queue was deleted */
+  QueueDelete = 'delete',
+  /** Queue is depleted */
+  QueueDepleted = 'depleted',
+  /** Message was queued */
+  QueueMessage = 'message',
+  /** Queue is ready to receive new messages */
+  QueueReady = 'ready',
+  /** Queue is saturated, i.e. max capacity was reached */
+  QueueSaturated = 'saturated',
+}
+
 export interface Queue {
   name: string;
   options: queueOptions;
@@ -48,25 +67,23 @@ export interface Queue {
   evictFirst(compareMessage?: Message): boolean;
   consume(onMessage: onMessage, consumeOptions?: consumeOptions, owner?: any): Consumer;
   assertConsumer(onMessage: onMessage, consumeOptions?: consumeOptions, owner?: any): Consumer;
-  get(options?: consumeOptions): Message;
+  get(options?: consumeOptions): Message | undefined;
   ack(message: Message, allUpTo?: boolean): void;
   nack(message: Message, allUpTo?: boolean, requeue?: boolean): void;
   reject(message: Message, requeue?: boolean): void;
   ackAll(): void;
   nackAll(requeue?: boolean): void;
-  peek(ignoreDelivered?: boolean): Message;
-  cancel(consumerTag: string, requeue?: boolean): void;
+  peek(ignoreDelivered?: boolean): Message | undefined;
+  cancel(consumerTag: string, requeue?: boolean): boolean;
   dismiss(onMessage: onMessage, requeue?: boolean): void;
   unbindConsumer(consumer: Consumer, requeue?: boolean): void;
   emit(eventName: string, content?: any): void;
-  on(eventName: string, handler: CallableFunction): Consumer;
-  off(eventName: string, handler: CallableFunction): Consumer;
+  on(eventName: string | QueueEventNames, handler: CallableFunction, options?: consumeOptions): Consumer;
+  off(eventName: string | QueueEventNames, handler: CallableFunction | consumeOptions): Consumer;
   purge(): number;
   getState(): QueueState;
   recover(state?: QueueState): Queue;
-  delete(options?: deleteQueueOptions): {
-    messageCount: number;
-  };
+  delete(options?: deleteQueueOptions): { messageCount: number };
   close(): void;
   stop(): void;
 }